Update mac install
Oh man, this was tricky.  I kept getting errors that said my compiler couldn't produce executables.  But only for chroma.

I tried reinstalling open-mpi with almost every combination available on homebrew until I found the `--without-fortran` flag to be the key.

But, even with that reinstalled, I still didn't succeed in compiling chroma.  The issue turned out to be the updated `LD_FLAGS[chroma]` defaults, which included `-Wl,zmuldefs`.  Getting rid of that on the mac cured the issue.