Compilation at FZJ
Successful compilation of chroma on JUQUEEN (without qphix).  The issue
was a mix of hdf5/1.8.15_BGQ and hdf5/1.8.15_bgclang modules.  Once the
paths were straightened out, the problem vanished.

I am having trouble with qphix due to an allegedly-missing libxml/xmlmemory.h.  Not sure WTF is going on there.  Could be that the bgq branches are too far behind the actively developed branches.

Move of jureca to fzj/jureca, which I think will be a long-term
organizational change.  I may leave the LLNL machines alone for now.
I'll add a usage mode to look around for machine files.