llnl/cab
Aztec is going out of service on 2016-01-04, so it makes sense to copy the scripts
to the machine cab now.  I have compiled them and they work.

This takes care of the reorganization I wanted for the LLNL aztec installations,
which can be deleted when they're out of use.

I also made minor fixes to usqcd.sh: re-fixing the shebang, and unquoting the message hooks.