Enable tab completion for machine files.
Previously, the first argument to usqcd.sh was the path to the machine file, but omitting the
leading machines/ and the trailing .sh.

Now, usqcd.sh immediately checks for and drops those strings from the machine file argument
(maintaining backward  compatibility) but accepting the full (relative) path to the machine file.
This allows the user to use their shell's tab completion feature to select a machine file.