Repair chroma on surface
Something must have happened with chroma where compiling `--with-quda` automatically engages the
quda multigrid solver, which previously wasn't found.

The solution is to compile the quda feature/multigrid branch (and enable the mg solver).

This puts us OFF the development branch on surface, so we should pay attention to whether there are
any fixes or big improvements over there.